# reference-apps/shared/openapi.yaml
openapi: 3.1.0
info:
  title: Colima Services - Reference API
  description: |
    Reference implementation demonstrating infrastructure integration patterns
    for production-grade applications.

    This API showcases:
    - HashiCorp Vault integration for secrets management
    - Multi-database connectivity (PostgreSQL, MySQL, MongoDB)
    - Redis caching and cluster management
    - RabbitMQ messaging patterns
    - Comprehensive health checks
    - Prometheus metrics integration

    **Note:** This is a reference implementation for learning and development.
    Production deployments should add authentication, rate limiting, and other
    security controls as documented in SECURITY_ASSESSMENT.md.

  version: 1.0.0
  contact:
    name: Development Team
    url: https://github.com/your-org/colima-services
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Code-First Implementation (FastAPI with decorators)
  - url: http://localhost:8001
    description: API-First Implementation (Generated from OpenAPI spec)
  - url: https://localhost:8443
    description: Code-First Implementation (HTTPS/TLS with Vault PKI)
  - url: https://localhost:8444
    description: API-First Implementation (HTTPS/TLS with Vault PKI)
...
paths:
...
  /redis/nodes/{node_name}/info:
    get:
      tags:
        - Redis Cluster
      summary: Get Node Info
      description: |-
        Get detailed information about a specific Redis node

        Args:
            node_name: Name of the node (redis-1, redis-2, redis-3)

        Returns detailed server information for the specified node
      operationId: get_node_info_redis_nodes__node_name__info_get
      parameters:
        - name: node_name
          in: path
          required: true
          schema:
            type: string
            title: Node Name
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /examples/cache/{key}:
    get:
      tags:
        - Cache Examples
      summary: Get Cache Value
      description: 'Example: Get a value from cache'
      operationId: get_cache_value_examples_cache__key__get
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
            pattern: ^[a-zA-Z0-9_:.-]+$
            description: 'Cache key (alphanumeric and: - _ : . only)'
            title: Key
          description: 'Cache key (alphanumeric and: - _ : . only)'
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheGetResponse'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
        - Cache Examples
      summary: Set Cache Value
      description: 'Example: Set a value in cache with optional TTL'
      operationId: set_cache_value_examples_cache__key__post
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
            pattern: ^[a-zA-Z0-9_:.-]+$
            description: 'Cache key (alphanumeric and: - _ : . only)'
            title: Key
          description: 'Cache key (alphanumeric and: - _ : . only)'
        - name: value
          in: query
          required: true
          schema:
            type: string
            minLength: 0
            maxLength: 10000
            description: Value to cache (max 10KB)
            title: Value
          description: Value to cache (max 10KB)
        - name: ttl
          in: query
          required: false
          schema:
            type: integer
            maximum: 86400
            minimum: 1
            description: Time to live in seconds (1s to 24h)
            title: Ttl
          description: Time to live in seconds (1s to 24h)
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheSetResponse'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - Cache Examples
      summary: Delete Cache Value
      description: 'Example: Delete a value from cache'
      operationId: delete_cache_value_examples_cache__key__delete
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
            pattern: ^[a-zA-Z0-9_:.-]+$
            description: 'Cache key (alphanumeric and: - _ : . only)'
            title: Key
          description: 'Cache key (alphanumeric and: - _ : . only)'
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheDeleteResponse'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    CacheDeleteResponse:
      properties:
        key:
          type: string
          title: Key
          description: Cache key
        deleted:
          type: boolean
          title: Deleted
          description: Whether key was deleted
        action:
          type: string
          title: Action
          description: Action performed
          default: delete
      type: object
      required:
        - key
        - deleted
      title: CacheDeleteResponse
      description: Response for cache delete operations
    CacheGetResponse:
      properties:
        key:
          type: string
          title: Key
          description: Cache key
        value:
          anyOf:
            - type: string
            - type: "null"
          title: Value
          description: Cached value (null if not found)
        exists:
          type: boolean
          title: Exists
          description: Whether key exists in cache
        ttl:
          anyOf:
            - type: integer
            - type: string
            - type: "null"
          title: Ttl
          description: Time to live in seconds or 'no expiration'
      type: object
      required:
        - key
        - exists
      title: CacheGetResponse
      description: Response for cache get operations
    CacheSetResponse:
      properties:
        key:
          type: string
          title: Key
          description: Cache key
        value:
          type: string
          title: Value
          description: Value that was set
        ttl:
          anyOf:
            - type: integer
            - type: "null"
          title: Ttl
          description: Time to live in seconds (null if no expiration)
        action:
          type: string
          title: Action
          description: Action performed
          default: set
      type: object
      required:
        - key
        - value
      title: CacheSetResponse
      description: Response for cache set operations
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API Key authentication (not currently implemented in reference app).
        Production deployments should implement API key validation.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication (not currently implemented in reference app).
        Production deployments should implement JWT validation with proper claims.